% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_close_tags.R
\name{get_close_tags}
\alias{get_close_tags}
\title{Get tags close to a series of locations}
\usage{
get_close_tags(
  pbfdir,
  lat,
  long,
  layers = DEFAULT_LAYERS,
  tags = DEFAULT_TAGS,
  dst = units::set_units(100, m),
  refilter_data = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{pbfdir}{File directory with .osm.pbf files containing map data for the relevant area. During processing the function will create a .p}

\item{lat}{numeric vector of latitudes}

\item{long}{numeric vector of longitudes}

\item{layers}{relevant layers from the pbf files (Do not touch if not familiar with OpenStreetMap data)}

\item{tags}{location tags of interest (Do not touch if not familiar with OpenStreetMap data)}

\item{dst}{max distance from each location to look for tags}

\item{refilter_data}{forcefully ignore cache}

\item{quiet}{less verbose function output}
}
\value{
list of vectors of tags, one for each location
}
\description{
\code{get_close_tags()} returns relevant tags which are at a certain distance from each location, giving a
meaningful sense of the type of place the person is in without providing exact data. No data is sent
online. If repeated queries are performed from the same directory with the same tags and the same or smaller distance,
most of the analysis will be drawn from cache and analysis time will be greatly reduced.
}
\examples{
mydir <- withr::local_tempdir()
file.copy(system.file("extdata", "tokelau.osm.pbf", package = "privlocR"), mydir)
get_close_tags(mydir, c(-9.1979860, -9.192079), c(-171.8501176, -171.856883))
get_close_tags(mydir, c(-9.1979860, -9.192079), c(-171.8501176, -171.856883), dst = units::set_units(10,m))
}
